add user registration, connect users with their contacts

+++ services/fetch/index.js
@@ -1,29 +1,53 @@
 const Contact = require('../schemas/contact')
+const User = require('../schemas/user')
 
-const getAllContacts = async () => {
-  return Contact.find()
+const getContacts = async (userId, query = {}) => {
+  const { docs: contacts, totalDocs: total } = await Contact.paginate(
+    { ...query, owner: userId },
+    {}
+  )
+  return { contacts, total }
 }
 
-const getContactById = async (id) => {
-  return Contact.findOne({ _id: id })
+const getContactById = async (userId, id) => {
+  return Contact.findOne({ _id: id, owner: userId })
 }
 
-const addContact = async (body) => {
-  return Contact.create({ ...body })
+const addContact = async (userId, body) => {
+  return Contact.create({ ...body, owner: userId })
 }
 
-const updateContact = async (id, body) => {
-  return Contact.findByIdAndUpdate({ _id: id }, body, { new: true })
+const updateContact = async (userId, id, body) => {
+  return Contact.findByIdAndUpdate({ _id: id, owner: userId }, body, {
+    new: true,
+  })
 }
 
-const removeContact = async (id) => {
-  return Contact.findByIdAndRemove({ _id: id })
+const removeContact = async (userId, id) => {
+  return Contact.findByIdAndRemove({ _id: id, owner: userId })
+}
+
+const findUser = async (filter) => {
+  return User.findOne(filter)
+}
+
+const addNewUser = ({ email, password }) => {
+  const newUser = new User({ email })
+  newUser.setPassword(password)
+  return newUser.save()
+}
+
+const updateUserById = (id, updateInfo) => {
+  return User.findByIdAndUpdate(id, updateInfo, { new: true })
 }
 
 module.exports = {
-  getAllContacts,
+  getContacts,
   getContactById,
   addContact,
   updateContact,
   removeContact,
+  findUser,
+  addNewUser,
+  updateUserById,
 }
